LayoutTests/http/tests/workers/service/registration-task-queue-scheduling-1-expected.txt

ALERT: Unexpected error received from server: UnknownError: Worker script successfully started, but it has no way to communicate yet
